sort of support multiple plugin instances
this lets one do

[plugin:geoip]
path = GeoIP.dat

[plugin:geoipcity]
clone = geoip
path = GeoIPCity.dat

it works in the API but the cli tool and web interface get confused and
show multiple identical entries for the plugin.  This is because
it uses .name from the plugin class, not the instance.

Fixing this will probably involve wrapping the plugins, which may help
other things too